==> 03_Inserting_the_data.ipynb <==
### Importing the Libraries

import pandas as pd
import psycopg2
from psycopg2 import sql

### Connecting to the database

DB_NAME = "Digital_Music_Store_Analysis"
DB_USER = "postgres"
DB_PASSWORD = "Ndsc@9652"
DB_HOST = "localhost"  
DB_PORT = "5432"

conn = psycopg2.connect(
    "postgresql://digitalmusicstoreanalysis_c7kr_user:ZY1ux0SCrWWhl94QTKPRVcqWSzUJb9F4@dpg-d07fdq9r0fns738gro70-a.oregon-postgres.render.com/digitalmusicstoreanalysis_c7kr"
)
cursor = conn.cursor()

def copy_csv_to_db(table_name, columns, file_path):
    try:
        with open(file_path, 'r') as f:
            cursor.copy_expert(
                sql.SQL("COPY {} ({}) FROM STDIN WITH CSV HEADER").format(
                    sql.Identifier(table_name),
                    sql.SQL(', ').join(map(sql.Identifier, columns))
                ),
                f
            )
        conn.commit()
        print(f"✅ Data inserted into {table_name} successfully!")
    except Exception as e:
        conn.rollback()
        print(f"❌ Error inserting data into {table_name}: {e}")


tables = [
    ("artist", "Datasets/Artist.csv"),
    ("album", "Datasets/Album.csv"),
    ("genre", "Datasets/Genre.csv"),
    ("mediatype", "Datasets/MediaType.csv"),
    ("track", "Datasets/Track.csv"),
    ("playlist", "Datasets/Playlist.csv"),
    ("playlisttrack", "Datasets/PlaylistTrack.csv"),
    ("employee", "Datasets/Employee.csv"),
    ("customer", "Datasets/Customer.csv"),
    ("invoice", "Datasets/Invoice.csv"),
    ("subscription", "Datasets/Subscription.csv"),
    ("useractivity", "Datasets/UserActivity.csv"),
    ("listeninghistory", "Datasets/ListeningHistory.csv"),
    ("adimpressions", "Datasets/AdImpressions.csv"),
    ("audiofeatures", "Datasets/AudioFeatures.csv"),
    ("artisttrack", "Datasets/ArtistTrack.csv"),
    ("artistalbum", "Datasets/ArtistAlbum.csv"),
    ("artistanalytics", "Datasets/ArtistAnalytics.csv")
]

tables_columns = {
    'artist': ["name"],
    'album': ["title", "artistid"],
    'genre': ["name"],
    'mediatype': ["name"],
    'track': ["name", "albumid", "mediatypeid", "genreid", "composer", "milliseconds", "bytes"],
    'playlist': ["name"],
    'playlisttrack': ["playlistid", "trackid"],
    'employee': ["lastname", "firstname", "title", "reportsto", "birthdate", "hiredate", "address", "city", "state", "country", "postalcode", "phone", "fax", "email"],
    'customer': ["firstname", "lastname", "company", "address", "city", "state", "country", "postalcode", "phone", "fax", "email", "supportrepid"],
    'invoice': ["customerid", "invoicedate", "billingaddress", "billingcity", "billingstate", "billingcountry", "billingpostalcode", "total"],
    'subscription': ["customerid", "plantype", "startdate", "enddate", "billingamount", "autorenew"],
    'useractivity': ["customerid", "activitytype", "timestamp"],
    'listeninghistory': ["customerid", "trackid", "listendate", "device"],
    'adimpressions': ["customerid", "adtype", "revenuegenerated", "timestamp"],
    'audiofeatures': ["trackid", "tempo", "keysignature", "danceability"],
    'artisttrack': ["artistid", "trackid"],
    'artistalbum': ["artistid", "albumid"],
    'artistanalytics': ["artistid", "totalstreams", "uniquelisteners", "revenueearned"]
}


for i in tables:
    copy_csv_to_db(i[0],tables_columns[i[0]],i[1])

conn.commit()
conn.close()
